Check for valid parameters

Check for valid parameters

check_params(params)

Arguments

params
data.frame or list of parameters

Value

Returns the aquifer type, depending on whether Dxx is specified ("confined") or PHIxx is specified ("unconfined")

Details

To evaluate the game, a data.frame or list of paramers must be supplied to evaluate_treaty or evaluate_treaty_cases. Generally speaking, these parameters should be non-negative. The parameters must include, for all scenarios:

Demand

Qf, Qs

Unit cost of alternative supply

p0f, p0s

Unit cost of pumping

B

Unit cost of recharge, Swiss

crs

Volume of recharge, with (T) and without (N) a treaty

rmT, rmN

Trust between players

gs, gf

Cost of the treaty

es, ef

And additional parameters depending on the type of aquifer. For confined aquifers:

Drawdown relationships

Dff, Dss, Dsf, Dfs

Groundwater depth without pumping

d0s, d0f

Recharge relationship with a treaty

DsrT, DfrT

Recharge relationship without a treaty

DsrN, DfrN

And for unconfined aquifers:

Drawdown relationships

PHIff, PHIss, PHIsf, PHIfs

Groundwater depth without pumping

dBs, dBf, h0s, h0f

Recharge relationship with a treaty

PHIsrT, PHIfrT

Recharge relationship without a treaty

PHIsrN, PHIfrN

Nonlinearity of the cost function

l

Examples

library(genevoisgame)
check_params(example_params_confined)

#> [1] "confined"

check_params(example_params_unconfined)

#> [1] "unconfined"

if (FALSE) check_params(example_params_confined[,-ncol(example_params_confined):-1])

params <- data.frame(example_params_confined,gs=c(0.2,0.8))
check_params(params)

#> [1] "confined"

if (FALSE) check_params(params[,-ncol(example_params_confined):-1])