Evaluate multiple treaty scenarios

Evaluate whether or not the treaty will be made under multiple scenarios

evaluate_treaty_cases(params_df, return_criteria = "qp",
  progress_bar = FALSE)

Arguments

params_df
Data.frame of parameters with each row sent to evaluate_treaty().

return_criteria
Character string containing letters that indicate output variables. See details.

progress_bar
Show a progress bar. Useful in unconfined aquifers for large N

Value

Returns a tibble containing z-values needed for the treaty and whether or not there is a treaty (i.e., if zRange > 0), as well as output specified by return_criteria.

If the aquifer in the game is confined (see ?check_params for how aquifer type is determined) the solution is exact. If the aquifer is unconfined, the solution is calculated numerically using multiroot from the rootSolve package, using default tolerance parameters. This means that there could be small errors in the results for zMinSwiss and zMaxFrench. For unconfined aquifers, zRange is rounded to the 6th decimal place to minimize the effect of numerical errors on the treaty outcome. If necessary, it can be recalculated as zRange = zMaxFrench - zMinSwiss.

In unconfined aquifers, it is possible that the game converges on a result that allows the aquifer to be fully depleted for as least one of the players (hi < 0). In this case, the results will included three additional columns: AD_fb,AD_nash,AD_cheat, representing logical values that indicate in which scenario the aquifer was depleted (first best, nash, or cheat). In the nonlinear game, this *should* only happen in one of two scenarios:

A treaty is signed to maximize join utility, but a cheat pumps more and the aquifer is depleted for the other player. In this case, the treaty is set to "N", even if trust is equal to 1.

The numerical root finder jumps to a value where the aquifer is fully depleted. This is unlikely, as the initial guesses are set to minimize the change of this occurring.

Details

Evaluate the treaty given multiple combinations of social, economic, and geophysical parameters. This function takes a data.frame of parameters, evaluates each row to see if a treaty is signed, and returns a tibble with the results and original params.

The parameter return_criteria can contain the following letters:

p - will return only parameters different from default. Otherwise all parameters returned

a - return all parameters (i.e., it's redundant to include a AND p)

u - return utilities of each player

d - return depth to water table for each player

Note that the cost of pumping is linear for confined aquifers, and nonlinear for unconfined aquifers, with the nonlinearity depending on the parameter l, with the key feature that the cost become infinite as h -> 0. For l = 0, the cost function is exponential and crosses 0 when h_i = dB_i. As l increases towards 1, the solution becomes more linear when the water table is near the surface. When l == 1, a linear solution is used and the cost remains linear even when the aquifer is fully depleted

Examples

library(genevoisgame)
evaluate_treaty_cases(rbind(example_params_confined,example_params_confined))

#> # A tibble: 2 x 10
#>   treaty zRange zMinSwiss zMaxFrench qshat qsstar qsdouble qfhat qfstar qfdouble
#>   <chr>   <dbl>     <dbl>      <dbl> <dbl>  <dbl>    <dbl> <dbl>  <dbl>    <dbl>
#> 1 Y      0.0443   -0.0222     0.0222  6.87   8.74     9.04  6.87   8.74     9.04
#> 2 Y      0.0443   -0.0222     0.0222  6.87   8.74     9.04  6.87   8.74     9.04

evaluate_treaty_cases(rbind(example_params_unconfined,example_params_unconfined),"qudp")

#> # A tibble: 2 x 26
#>   treaty  zRange zMinSwiss zMaxFrench qshat qsstar qsdouble qfhat qfstar
#>   <chr>    <dbl>     <dbl>      <dbl> <dbl>  <dbl>    <dbl> <dbl>  <dbl>
#> 1 Y      0.00954   -0.0236    -0.0141  5.98   6.85     6.96  5.09   6.10
#> 2 Y      0.00954   -0.0236    -0.0141  5.98   6.85     6.96  5.09   6.10
#> # … with 17 more variables: qfdouble <dbl>, Us_hat <dbl>, Us_star <dbl>,
#> #   Us_double <dbl>, Us_hat_double <dbl>, Uf_hat <dbl>, Uf_star <dbl>,
#> #   Uf_double <dbl>, Uf_hat_double <dbl>, ds_hat <dbl>, ds_star <dbl>,
#> #   ds_double <dbl>, ds_hat_double <dbl>, df_hat <dbl>, df_star <dbl>,
#> #   df_double <dbl>, df_hat_double <dbl>